V4 support: initial setup
Still a bit left to do, but I'd like some feedback on the structure and
some of the pending changes
