+++ src/keystore.rs
@@ -1,40 +1,44 @@
-use hex::{FromHex, ToHex};
-use serde::{de::Deserializer, ser::Serializer, Deserialize, Serialize};
-use uuid::Uuid;
+use serde::{Deserialize, Deserializer, Serialize};
 
-#[cfg(feature = "geth-compat")]
-use ethereum_types::H160 as Address;
+use crate::{v3, v4};
 
-#[derive(Debug, Deserialize, Serialize)]
-/// This struct represents the deserialized form of an encrypted JSON keystore based on the
-/// [Web3 Secret Storage Definition](https://github.com/ethereum/wiki/wiki/Web3-Secret-Storage-Definition).
-pub struct EthKeystore {
-    #[cfg(feature = "geth-compat")]
-    pub address: Address,
+#[derive(Debug, Serialize)]
+#[serde(tag = "version")]
+pub enum EthKeystore {
+    #[serde(rename = "3")]
+    V3(v3::EthKeystoreV3),
 
-    pub crypto: CryptoJson,
-    pub id: Uuid,
-    pub version: u8,
+    // TODO: add a v4 feature flag
+    #[serde(rename = "4")]
+    V4(v4::EthKeystoreV4),
 }
 
-#[derive(Debug, Deserialize, Serialize)]
-/// Represents the "crypto" part of an encrypted JSON keystore.
-pub struct CryptoJson {
-    pub cipher: String,
-    pub cipherparams: CipherparamsJson,
-    #[serde(serialize_with = "buffer_to_hex", deserialize_with = "hex_to_buffer")]
-    pub ciphertext: Vec<u8>,
-    pub kdf: KdfType,
-    pub kdfparams: KdfparamsType,
-    #[serde(serialize_with = "buffer_to_hex", deserialize_with = "hex_to_buffer")]
-    pub mac: Vec<u8>,
-}
+impl<'de> Deserialize<'de> for EthKeystore {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let v = serde_json::Value::deserialize(deserializer)?;
 
-#[derive(Debug, Deserialize, Serialize)]
-/// Represents the "cipherparams" part of an encrypted JSON keystore.
-pub struct CipherparamsJson {
-    #[serde(serialize_with = "buffer_to_hex", deserialize_with = "hex_to_buffer")]
-    pub iv: Vec<u8>,
+        match v["version"].as_i64() {
+            // TODO tests/test-key/my_keystore has no version field
+            // should we default to v3 to preserve compatibility, or raise an error?
+            Some(3) | None => {
+                let v3 = serde_json::from_value::<v3::EthKeystoreV3>(v)
+                    .map_err(|_| serde::de::Error::custom("asd"))?;
+                Ok(EthKeystore::V3(v3))
+            }
+            Some(4) => {
+                let v4 = serde_json::from_value::<v4::EthKeystoreV4>(v)
+                    .map_err(|_| serde::de::Error::custom("asd"))?;
+                Ok(EthKeystore::V4(v4))
+            }
+            v => Err(serde::de::Error::custom(format!(
+                "Unsupported version: {}",
+                v.unwrap_or_default()
+            ))),
+        }
+    }
 }
 
 #[derive(Debug, Deserialize, Eq, PartialEq, Serialize)]
@@ -45,79 +49,14 @@ pub enum KdfType {
     Scrypt,
 }
 
-#[derive(Debug, Deserialize, Eq, PartialEq, Serialize)]
-#[serde(untagged)]
-/// Defines the various parameters used in the supported KDFs.
-pub enum KdfparamsType {
-    Pbkdf2 {
-        c: u32,
-        dklen: u8,
-        prf: String,
-        #[serde(serialize_with = "buffer_to_hex", deserialize_with = "hex_to_buffer")]
-        salt: Vec<u8>,
-    },
-    Scrypt {
-        dklen: u8,
-        n: u32,
-        p: u32,
-        r: u32,
-        #[serde(serialize_with = "buffer_to_hex", deserialize_with = "hex_to_buffer")]
-        salt: Vec<u8>,
-    },
-}
-
-fn buffer_to_hex<T, S>(buffer: &T, serializer: S) -> Result<S::Ok, S::Error>
-where
-    T: AsRef<[u8]>,
-    S: Serializer,
-{
-    serializer.serialize_str(&buffer.encode_hex::<String>())
-}
-
-fn hex_to_buffer<'de, D>(deserializer: D) -> Result<Vec<u8>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    use serde::de::Error;
-    String::deserialize(deserializer)
-        .and_then(|string| Vec::from_hex(&string).map_err(|err| Error::custom(err.to_string())))
-}
-
 #[cfg(test)]
 mod tests {
-    use super::*;
 
-    #[cfg(feature = "geth-compat")]
-    #[test]
-    fn deserialize_geth_compat_keystore() {
-        let data = r#"
-        {
-            "address": "00000398232e2064f896018496b4b44b3d62751f",
-            "crypto": {
-                "cipher": "aes-128-ctr",
-                "ciphertext": "4f784cd629a7caf34b488e36fb96aad8a8f943a6ce31c7deab950c5e3a5b1c43",
-                "cipherparams": {
-                    "iv": "76f07196b3c94f25b8f34d869493f640"
-                },
-                "kdf": "scrypt",
-                "kdfparams": {
-                    "dklen": 32,
-                    "n": 262144,
-                    "p": 1,
-                    "r": 8,
-                    "salt": "1e7be4ce8351dd1710b0885438414b1748a81f1af510eda11e4d1f99c8d43975"
-                },
-                "mac": "5b5433575a2418c1c813337a88b4099baa2f534e5dabeba86979d538c1f594d8"
-            },
-            "id": "6c4485f3-3cc0-4081-848e-8bf489f2c262",
-            "version": 3
-        }"#;
-        let keystore: EthKeystore = serde_json::from_str(data).unwrap();
-        assert_eq!(
-            keystore.address.as_bytes().to_vec(),
-            hex::decode("00000398232e2064f896018496b4b44b3d62751f").unwrap()
-        );
-    }
+    use hex::FromHex;
+    use uuid::Uuid;
+
+    use crate::keystore::EthKeystore;
+    use crate::v3::{KdfType, KdfparamsType};
 
     #[cfg(not(feature = "geth-compat"))]
     #[test]
@@ -143,99 +82,58 @@ mod tests {
             "version" : 3
         }"#;
         let keystore: EthKeystore = serde_json::from_str(data).unwrap();
-        assert_eq!(keystore.version, 3);
-        assert_eq!(
-            keystore.id,
-            Uuid::parse_str("3198bc9c-6672-5ab3-d995-4942343ae5b6").unwrap()
-        );
-        assert_eq!(keystore.crypto.cipher, "aes-128-ctr");
-        assert_eq!(
-            keystore.crypto.cipherparams.iv,
-            Vec::from_hex("6087dab2f9fdbbfaddc31a909735c1e6").unwrap()
-        );
-        assert_eq!(
-            keystore.crypto.ciphertext,
-            Vec::from_hex("5318b4d5bcd28de64ee5559e671353e16f075ecae9f99c7a79a38af5f869aa46")
-                .unwrap()
-        );
-        assert_eq!(keystore.crypto.kdf, KdfType::Pbkdf2);
-        assert_eq!(
-            keystore.crypto.kdfparams,
-            KdfparamsType::Pbkdf2 {
-                c: 262144,
-                dklen: 32,
-                prf: String::from("hmac-sha256"),
-                salt: Vec::from_hex(
-                    "ae3cd4e7013836a3df6bd7241b12db061dbe2c6785853cce422d148a624ce0bd"
-                )
-                .unwrap(),
-            }
-        );
-        assert_eq!(
-            keystore.crypto.mac,
-            Vec::from_hex("517ead924a9d0dc3124507e3393d175ce3ff7c1e96529c6c555ce9e51205e9b2")
-                .unwrap()
-        );
+
+        match keystore {
+            EthKeystore::V3(_) => (),
+            _ => panic!("not v3"), // TODO: how to properly fail here?
+        };
     }
 
     #[cfg(not(feature = "geth-compat"))]
     #[test]
     fn test_deserialize_scrypt() {
+        // Test vec from: https://eips.ethereum.org/EIPS/eip-2335
+
+        use hex::FromHex;
         let data = r#"
         {
-            "crypto" : {
-                "cipher" : "aes-128-ctr",
-                "cipherparams" : {
-                    "iv" : "83dbcc02d8ccb40e466191a123791e0e"
+            "crypto": {
+                "kdf": {
+                    "function": "scrypt",
+                    "params": {
+                        "dklen": 32,
+                        "n": 262144,
+                        "p": 1,
+                        "r": 8,
+                        "salt": "d4e56740f876aef8c010b86a40d5f56745a118d0906a34e69aec8c0db1cb8fa3"
+                    },
+                    "message": ""
                 },
-                "ciphertext" : "d172bf743a674da9cdad04534d56926ef8358534d458fffccd4e6ad2fbde479c",
-                "kdf" : "scrypt",
-                "kdfparams" : {
-                    "dklen" : 32,
-                    "n" : 262144,
-                    "p" : 8,
-                    "r" : 1,
-                    "salt" : "ab0c7876052600dd703518d6fc3fe8984592145b591fc8fb5c6d43190334ba19"
+                "checksum": {
+                    "function": "sha256",
+                    "params": {},
+                    "message": "d2217fe5f3e9a1e34581ef8a78f7c9928e436d36dacc5e846690a5581e8ea484"
                 },
-                "mac" : "2103ac29920d71da29f15d75b4a16dbe95cfd7ff8faea1056c33131d846e3097"
+                "cipher": {
+                    "function": "aes-128-ctr",
+                    "params": {
+                        "iv": "264daa3f303d7259501c93d997d84fe6"
+                    },
+                    "message": "06ae90d55fe0a6e9c5c3bc5b170827b2e5cce3929ed3f116c2811e6366dfe20f"
+                }
             },
-            "id" : "3198bc9c-6672-5ab3-d995-4942343ae5b6",
-            "version" : 3
+            "description": "This is a test keystore that uses scrypt to secure the secret.",
+            "pubkey": "9612d7a727c9d0a22e185a1c768478dfe919cada9266988cb32359c11f2b7b27f4ae4040902382ae2910c15e2b420d07",
+            "path": "m/12381/60/3141592653/589793238",
+            "uuid": "1d85ae20-35c5-4611-98e8-aa14a633906f",
+            "version": 4
         }"#;
+
         let keystore: EthKeystore = serde_json::from_str(data).unwrap();
-        assert_eq!(keystore.version, 3);
-        assert_eq!(
-            keystore.id,
-            Uuid::parse_str("3198bc9c-6672-5ab3-d995-4942343ae5b6").unwrap()
-        );
-        assert_eq!(keystore.crypto.cipher, "aes-128-ctr");
-        assert_eq!(
-            keystore.crypto.cipherparams.iv,
-            Vec::from_hex("83dbcc02d8ccb40e466191a123791e0e").unwrap()
-        );
-        assert_eq!(
-            keystore.crypto.ciphertext,
-            Vec::from_hex("d172bf743a674da9cdad04534d56926ef8358534d458fffccd4e6ad2fbde479c")
-                .unwrap()
-        );
-        assert_eq!(keystore.crypto.kdf, KdfType::Scrypt);
-        assert_eq!(
-            keystore.crypto.kdfparams,
-            KdfparamsType::Scrypt {
-                dklen: 32,
-                n: 262144,
-                p: 8,
-                r: 1,
-                salt: Vec::from_hex(
-                    "ab0c7876052600dd703518d6fc3fe8984592145b591fc8fb5c6d43190334ba19"
-                )
-                .unwrap(),
-            }
-        );
-        assert_eq!(
-            keystore.crypto.mac,
-            Vec::from_hex("2103ac29920d71da29f15d75b4a16dbe95cfd7ff8faea1056c33131d846e3097")
-                .unwrap()
-        );
+
+        match keystore {
+            EthKeystore::V4(_) => (),
+            _ => panic!("not v3"), // TODO: how to properly fail here?
+        };
     }
 }
