+++ src/utils.rs
@@ -1,3 +1,6 @@
+use hex::{FromHex, ToHex};
+use serde::{Deserialize, Deserializer, Serializer};
+
 #[cfg(feature = "geth-compat")]
 pub mod geth_compat {
     use ethereum_types::H160 as Address;
@@ -29,4 +32,21 @@ pub mod geth_compat {
         hasher.update(bytes.as_ref());
         hasher.finalize().into()
     }
+}
+
+pub(crate) fn buffer_to_hex<T, S>(buffer: &T, serializer: S) -> Result<S::Ok, S::Error>
+where
+    T: AsRef<[u8]>,
+    S: Serializer,
+{
+    serializer.serialize_str(&buffer.encode_hex::<String>())
+}
+
+pub(crate) fn hex_to_buffer<'de, D>(deserializer: D) -> Result<Vec<u8>, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    use serde::de::Error;
+    String::deserialize(deserializer)
+        .and_then(|string| Vec::from_hex(&string).map_err(|err| Error::custom(err.to_string())))
 }
